+++ src/api/mock.ts
@@ -6,7 +6,7 @@ import {Connection, Message} from './connection';
 import {QueryObserverManager, MESSAGE_ADDED, MESSAGE_CHANGED, MESSAGE_REMOVED} from './queryobserver';
 import {GenError} from '../core/errors/error';
 import {APIError} from './errors';
-import {Query, Sample, Collection, Data} from './types/rest';
+import {Query, Collection, Data} from './types/rest';
 import {ResolweApi} from './index';
 import {compose} from '../core/utils/lang';
 import * as random from '../core/utils/random';
@@ -735,15 +735,6 @@ export function orderingQuery<T>(query: Query, items: T[]): T[] {
     const orderingDirections = _.map(ordering, (column) => column[0] === '-' ? 'desc' : 'asc');
     const orderingColumns = _.map(ordering, (column) => column[0] === '-' ? column.substr(1) : column);
     return _.sortByOrder(items, orderingColumns, orderingDirections);
-}
-
-/**
- * Helper function for supporting filtering by descriptor_completed, which can be used as a [[MockQueryEvaluator]].
- */
-export function annotatedQuery<T extends Sample>(query: any, items: T[]): T[] {
-    if (_.isUndefined(query.descriptor_completed) || _.isNull(query.descriptor_completed)) return items;
-
-    return _.filter(items, (item) => item.descriptor_completed === query.descriptor_completed);
 }
 
 /**